Breaking change - renaming/adding parameters to match other functions

+++ functions/Update-PowerUpConfig.ps1
@@ -33,7 +33,7 @@
 			'ConnectionTimeout', 'ExecutionTimeout', 'Encrypt', 'Credential', 'Username',
 			'Password', 'SchemaVersionTable', 'Silent', 'Variables'
 		)]
-		[string]$Config,
+		[string]$ConfigName,
 		[Parameter(ParameterSetName = 'Value',
 			Mandatory = $true,
 			Position = 3 )]
@@ -41,12 +41,19 @@
 		[Parameter(ParameterSetName = 'Hashtable',
 			Mandatory = $true,
 			Position = 2 )]
-		[hashtable]$Values,
+		[Alias('Config')]
+		[hashtable]$Configuration,
 		[Parameter(ParameterSetName = 'File',
 			Mandatory = $true,
 			Position = 2 )]
 		[Alias('ConfigFile')]
 		[string]$ConfigurationFile,
+		[Parameter(ParameterSetName = 'Variables',
+			Mandatory = $true,
+			Position = 2 )]
+		[Parameter(ParameterSetName = 'Hashtable')]
+		[Parameter(ParameterSetName = 'File')]
+		[hashtable]$Variables,
 		[switch]$Unpacked
 	)
 	begin {
@@ -58,49 +65,60 @@
 				$workFolder = $pFile
 			}
 			else {
-				$workFolder = New-TempWorkspaceFolder
+				if ($pscmdlet.ShouldProcess([System.IO.Path]::GetTempPath(), "Creating new temporary folder")) {
+					$workFolder = New-TempWorkspaceFolder
+				}
+				else {
+					$workFolder = "NonExistingPath"
+				}
 			}
 			try {
+				$packageFile = [PowerUpConfig]::GetPackageFileName()
 				if (!$Unpacked) {
 					#Extract package files
-					Write-Verbose "Extracting package file from $pFile to $workFolder"
-					$packageFile = [PowerUpConfig]::GetPackageFileName()
-					Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item $packageFile
+					if ($pscmdlet.ShouldProcess($pFile, "Extracting package file to $workFolder")) {
+						Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item $packageFile
+					}
 				}
 			
-				Write-Verbose "Reading package file from $workFolder"
-				$package = [PowerUpPackage]::FromFile((Join-Path $workFolder $packageFile))
+				if ($pscmdlet.ShouldProcess($packageFile, "Reading package file from $workFolder")) {
+					$package = [PowerUpPackage]::FromFile((Join-Path $workFolder $packageFile))
+				}
+				else {
+					$package = [PowerUpPackage]::new()
+				}
 				$configFile = $package.ConfigurationFile
 
 				if (!$Unpacked) {
-					Write-Verbose "Extracting config file $configFile from $pFile"
-					Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item $configFile
+					if ($pscmdlet.ShouldProcess($configFile, "Extracting config file from $pFile")) {
+						Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item $configFile
+					}
 				}
-
-				Write-Verbose "Reading config file $configFile from $workFolder"
-				$configTempFile = (Join-Path $workFolder $configFile)
-				$configObject = [PowerUpConfig]::FromFile($configTempFile)
-
+				
 				#Assign new values
 				Write-Verbose "Assigning new values to the config"
 
 				if ($PSCmdlet.ParameterSetName -eq 'Value') {
-					$newConfig = @{ $Config = $Value }
+					$newConfig = @{ $ConfigName = $Value }
 				}
 				elseif ($PSCmdlet.ParameterSetName -eq 'Hashtable') {
-					$newConfig = $Values
+					$newConfig = $Configuration
 				}
 				elseif ($PSCmdlet.ParameterSetName -eq 'File') {
 					$newConfig = (Get-PowerUpConfig -Path $ConfigurationFile).AsHashtable()
 				}
-
-				Write-Verbose "Processing keys $($newConfig.Keys -join ', ')"
-				foreach ($property in $newConfig.Keys) {
-					$configObject.SetValue($property,$newConfig.$property)
+				#Overriding Variables
+				if ($Variables) {
+					if ($PSCmdlet.ParameterSetName -ne 'Variables') { $newConfig.Remove('Variables') }
+					$newConfig += @{ Variables = $Variables}
 				}
 
-				Write-Verbose "Saving the config file $configTempFile"
-				$configObject.SaveToFile($configTempFile, $true)
+				$configTempFile = (Join-Path $workFolder $configFile)
+				$configObject = Get-PowerUpConfig -Path $configTempFile -Configuration $newConfig
+
+				if ($pscmdlet.ShouldProcess($configTempFile, "Saving the config file")) {
+					$configObject.SaveToFile($configTempFile, $true)
+				}
 
 				if ($pscmdlet.ShouldProcess($pFile, "Updating package with a new config file")) {
 					$null = Add-ArchiveItem -Path $pFile -Item $configTempFile
@@ -111,8 +129,9 @@
 			}
 			finally {
 				if (!$Unpacked -and $workFolder.Name -like 'PowerUpWorkspace*') {
-					Write-Verbose "Removing temporary folder $workFolder"
-					Remove-Item $workFolder -Recurse -Force
+					if ($pscmdlet.ShouldProcess($workFolder, "Removing temporary folder")) {
+						Remove-Item $workFolder -Recurse -Force
+					}
 				}
 			}
 		}
